convert metro config and queryFiles utilities to TS
