--- scripts/configs/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  ...require('./configure-metro')
-};
